requirements.txt
Flask
qrcode
pyzbar
Pillow

